Extract keep specific code to `lib/keep.js`
